Add numbers and operators to displayable values
Numbers and operators can be pushed on the calculator.
Pressing will add the value to an array that keeps track of the order.
The display is update with more user friendly icons but the array keeps
track of mroe useful names for later calculation.
Also when trying to enter to operators back to back the screen shakes
red.