Add parentheses functionality
Left and right parentheses function as expected on a calculator.
A right para only can be enterd if there are more left than right.
Nested paras are possible as well.
When equal is pressed the calculate function will search for a
set of parentheses if it finds one it calls the calculate function
on the substring. Then continues till no paratheses are left and calls
the operator handler. It repeats until only a single value remains.