Add sqrt, factorial, ans, and decimal buttons
Sqrt and factorial are treated as functions that can be applied to
a number. They add parentheses when clicked too.
The ANS button just adds the previous calculation.
The decimal button acts as a number that cannot be entered into a
value twice.
Extra: parentheses must be closed off or invalidInput is called